requirements.txt
fastapi
uvicorn
requests
openai
python-dotenv
sse-starlette